Make window non-restorable and remove top buttons
Previously, through a combination of turning off permissions and using
certain gestures at the exact right moment, it was possible to make
KeyboardCleaner get stuck in an incorrect window size, with no way to
restore it to the original size. The window will now never restore to
the previous size. This issue only affected the permissions warning
message.